CVS-168 Fix
