+++ .env.sample
@@ -1,3 +1,6 @@
 ORIGIN_URI=1.2.3.4
-STORAGE_DIR=/var/lib/charv/storage
-SOCKETS_DIR=/tmp/charv
+IMAGES_DIR=/var/lib/charv/images/
+GUESTS_DIR=/var/lib/charv/guests/
+STORAGE_POOLS_DIR=/var/lib/charv/storage_pools/
+QMP_DIR=/tmp/charv/qmp/
+VNC_DIR=/tmp/charv/vnc/
