+++ infrastructure/utils/uuid.go
@@ -12,10 +12,5 @@ func NewID() models.ID {
 }
 
 func (i id) GenID() (uuid.UUID, error) {
-	id, err := uuid.NewRandom() // UUID Version 4
-	if err != nil {
-		return uuid.UUID{}, err
-	}
-
-	return id, nil
+	return uuid.NewRandom() // UUID Version 4
 }
